requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
docutils==0.21.2
duckduckgo_search==5.3.1b1
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.59.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-language==2.13.4
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
gTTS==2.5.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
jaraco.context==5.3.0
Jinja2==3.1.4
lockfile==0.12.2
MarkupSafe==2.1.5
more-itertools==10.3.0
multidict==6.0.5
numpy==2.0.0
openai==1.29.0
packaging==24.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytube==15.0.0
redis==5.0.7
regex==2024.5.15
requests==2.31.0
rsa==4.9
setuptools==70.3.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
spotipy==2.23.0
tenacity==8.3.0
tiktoken==0.7.0
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
vertexai==1.49.0
Werkzeug==3.0.3
whois==0.9.27
wolframalpha==5.0.1
xmltodict==0.13.0
yarl==1.9.4
